pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = "Simple MLOps project for MLOps course"
authors = ["Alexandr Korchemnyi <akorchemnyj@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.3.1"
torch = {version = "^2.1.0+cu118", source = "pytorch"}
torchvision = {version = "^0.16.0+cu118", source = "pytorch"}
lightning = "^2.1.2"
einops = "^0.7.0"
wandb = "^0.16.0"
six = "^1.16.0"
mlflow = "^2.8.1"
gitpython = "^3.1.40"
onnx = "^1.15.0"
fastapi = "^0.104.1"
onnxruntime = "^1.16.3"
uvicorn = "^0.24.0.post1"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
dvc = "^3.30.3"
dvc-gdrive = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90
target-version = ["py310"]
